+++ src/Client/ClientCore/Rendering/Scenes/Game/Gui/TemplateEditor/BlockTemplateEditorTab.cs
@@ -17,6 +17,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
+using System.Text;
 using ImGuiNET;
 using Sovereign.ClientCore.Components.Indexers;
 using Sovereign.ClientCore.Rendering.Gui;
@@ -24,6 +25,7 @@ using Sovereign.ClientCore.Rendering.Materials;
 using Sovereign.ClientCore.Rendering.Sprites.TileSprites;
 using Sovereign.EngineCore.Components;
 using Sovereign.EngineCore.Components.Types;
+using Sovereign.EngineCore.Components.Validators;
 using Sovereign.EngineCore.Entities;
 using Sovereign.EngineCore.Events;
 
@@ -39,8 +41,13 @@ public class BlockTemplateEditorTab
     /// </summary>
     private const uint SelectionColor = 0xFF773333;
 
+    private const float BottomOffsetY = 85;
+
+    private readonly EntityDefinitionGenerator definitionGenerator;
+
     private readonly EntityTable entityTable;
     private readonly IEventSender eventSender;
+    private readonly GuiComponentEditors guiComponentEditors;
 
     private readonly GuiExtensions guiExtensions;
     private readonly BlockTemplateEntityIndexer indexer;
@@ -48,8 +55,14 @@ public class BlockTemplateEditorTab
     private readonly MaterialManager materialManager;
     private readonly MaterialModifierComponentCollection materialModifiers;
     private readonly MaterialComponentCollection materials;
+    private readonly NameComponentValidator nameComponentValidator;
     private readonly NameComponentCollection names;
     private bool initialized;
+    private bool inputDrawable;
+    private int inputMaterial;
+    private int inputMaterialModifier;
+    private string inputName = "";
+    private EntityDefinition selectedDefinition = new();
     private ulong selectedEntityId;
     private int selectedIndex;
     private List<ulong> sortedTemplateEntityIds = new();
@@ -57,7 +70,9 @@ public class BlockTemplateEditorTab
     public BlockTemplateEditorTab(BlockTemplateEntityIndexer indexer, GuiExtensions guiExtensions,
         MaterialComponentCollection materials, MaterialModifierComponentCollection materialModifiers,
         NameComponentCollection names, MaterialManager materialManager, EntityTable entityTable,
-        TemplateEditorInternalController internalController, IEventSender eventSender)
+        TemplateEditorInternalController internalController, IEventSender eventSender,
+        EntityDefinitionGenerator definitionGenerator, NameComponentValidator nameComponentValidator,
+        GuiComponentEditors guiComponentEditors)
     {
         this.indexer = indexer;
         this.guiExtensions = guiExtensions;
@@ -68,6 +83,9 @@ public class BlockTemplateEditorTab
         this.entityTable = entityTable;
         this.internalController = internalController;
         this.eventSender = eventSender;
+        this.definitionGenerator = definitionGenerator;
+        this.nameComponentValidator = nameComponentValidator;
+        this.guiComponentEditors = guiComponentEditors;
 
         indexer.OnIndexModified += RefreshList;
     }
@@ -111,7 +129,7 @@ public class BlockTemplateEditorTab
         var maxSize = ImGui.GetWindowSize();
         if (ImGui.BeginTable("blockTemplateBrowser", 2,
                 ImGuiTableFlags.ScrollX | ImGuiTableFlags.ScrollY | ImGuiTableFlags.BordersOuter |
-                ImGuiTableFlags.RowBg, new Vector2 { X = 254.0f, Y = maxSize.Y - 90 }))
+                ImGuiTableFlags.RowBg, new Vector2 { X = 254.0f, Y = maxSize.Y - BottomOffsetY }))
         {
             ImGui.TableSetupColumn("Material");
             ImGui.TableSetupColumn("Name");
@@ -186,6 +204,85 @@ public class BlockTemplateEditorTab
     /// </summary>
     private void RenderEditor()
     {
+        if (selectedIndex < 0) return;
+
+        RenderComponentTable();
+        RenderEditorControls();
+    }
+
+    /// <summary>
+    ///     Renders the table of editable components for the block template.
+    /// </summary>
+    private void RenderComponentTable()
+    {
+        var maxSize = ImGui.GetWindowSize();
+        if (!ImGui.BeginTable("componentTable", 1,
+                ImGuiTableFlags.ScrollX | ImGuiTableFlags.ScrollY | ImGuiTableFlags.SizingStretchProp,
+                new Vector2 { X = maxSize.X - 276, Y = maxSize.Y - BottomOffsetY })) return;
+
+        ImGui.TableNextColumn();
+
+        if (ImGui.CollapsingHeader("Basic Information", ImGuiTreeNodeFlags.DefaultOpen))
+        {
+            if (ImGui.BeginTable("BasicInfo", 2, ImGuiTableFlags.SizingFixedFit))
+            {
+                ImGui.TableNextColumn();
+                ImGui.Text("Name:");
+                ImGui.TableNextColumn();
+                guiComponentEditors.NameEdit("##name", ref inputName);
+
+                ImGui.EndTable();
+            }
+        }
+
+        if (ImGui.CollapsingHeader("Appearance", ImGuiTreeNodeFlags.DefaultOpen))
+        {
+            if (ImGui.BeginTable("Appearance", 2, ImGuiTableFlags.SizingFixedFit))
+            {
+                ImGui.TableNextColumn();
+                ImGui.Text("Drawable:");
+                ImGui.TableNextColumn();
+                ImGui.Checkbox("##drawable", ref inputDrawable);
+
+                ImGui.EndTable();
+            }
+
+            guiComponentEditors.MaterialEdit(ref inputMaterial, ref inputMaterialModifier);
+        }
+
+        ImGui.EndTable();
+    }
+
+    /// <summary>
+    ///     Renders the save/cancel controls at the bottom of the editor.
+    /// </summary>
+    private void RenderEditorControls()
+    {
+        if (ImGui.BeginTable("Controls", 2, ImGuiTableFlags.SizingFixedFit))
+        {
+            ImGui.TableNextColumn();
+            var canSave = CanSave(out var reason);
+            if (!canSave) ImGui.BeginDisabled();
+            if (ImGui.Button("Save")) Save();
+            if (!canSave)
+            {
+                if (ImGui.IsItemHovered(ImGuiHoveredFlags.AllowWhenDisabled))
+                {
+                    if (ImGui.BeginTooltip())
+                    {
+                        ImGui.Text(reason);
+                        ImGui.EndTooltip();
+                    }
+                }
+
+                ImGui.EndDisabled();
+            }
+
+            ImGui.TableNextColumn();
+            if (ImGui.Button("Cancel")) Cancel();
+
+            ImGui.EndTable();
+        }
     }
 
     /// <summary>
@@ -205,6 +302,57 @@ public class BlockTemplateEditorTab
     }
 
     /// <summary>
+    ///     Validates the currently edited entity and determines if it can be saved.
+    /// </summary>
+    /// <param name="reason">If invalid, will be set to the reason(s) why.</param>
+    /// <returns>true if valid and can be saved, false otherwise.</returns>
+    private bool CanSave(out string reason)
+    {
+        var sb = new StringBuilder();
+        var valid = true;
+
+        if (!nameComponentValidator.IsValid(inputName))
+        {
+            sb.Append("Name is invalid.\n");
+            valid = false;
+        }
+
+        if (inputMaterial < 1 || inputMaterial >= materialManager.Materials.Count)
+        {
+            sb.Append("Material ID is invalid.\n");
+            valid = false;
+        }
+        else if (inputMaterialModifier < 0 ||
+                 inputMaterialModifier >= materialManager.Materials[inputMaterial].MaterialSubtypes.Count)
+        {
+            sb.Append("Material modifier is invalid.\n");
+            valid = false;
+        }
+
+        reason = sb.ToString().TrimEnd();
+        return valid;
+    }
+
+    /// <summary>
+    ///     Saves the currently edited template entity.
+    /// </summary>
+    private void Save()
+    {
+        selectedDefinition.Name = inputName.Length > 0 ? inputName : null;
+        selectedDefinition.Drawable = inputDrawable;
+        selectedDefinition.Material = new MaterialPair(inputMaterial, inputMaterialModifier);
+        SaveDefinition(selectedDefinition);
+    }
+
+    /// <summary>
+    ///     Cancels editing and reverts any changes to the currently selected entity.
+    /// </summary>
+    private void Cancel()
+    {
+        Select(selectedIndex);
+    }
+
+    /// <summary>
     ///     Saves a template entity to the server given its definition.
     /// </summary>
     /// <param name="definition">Definition.</param>
@@ -213,11 +361,25 @@ public class BlockTemplateEditorTab
         internalController.UpdateTemplateEntity(eventSender, definition);
     }
 
+    /// <summary>
+    ///     Selects the block template entity at the given position in the ordered list.
+    /// </summary>
+    /// <param name="index">Index into ordered list of block template entities.</param>
     private void Select(int index)
     {
         selectedIndex = index;
         if (selectedIndex >= 0 && selectedIndex < sortedTemplateEntityIds.Count)
+        {
             selectedEntityId = sortedTemplateEntityIds[index];
+            selectedDefinition = definitionGenerator.GenerateDefinition(selectedEntityId);
+            inputName = selectedDefinition.Name ?? "";
+            inputDrawable = selectedDefinition.Drawable;
+            if (selectedDefinition.Material != null)
+            {
+                inputMaterial = selectedDefinition.Material.MaterialId;
+                inputMaterialModifier = selectedDefinition.Material.MaterialModifier;
+            }
+        }
     }
 
     /// <summary>