+++ src/Client/ClientCore/Rendering/Scenes/Game/Gui/TemplateEditor/TemplateEditorGui.cs
@@ -37,8 +37,11 @@ public class TemplateEditorGui
     public void Render()
     {
         ImGui.SetNextWindowSize(new Vector2(900.0f, 640.0f), ImGuiCond.FirstUseEver);
+        ImGui.PushStyleVar(ImGuiStyleVar.WindowPadding, new Vector2(16.0f, 0.0f));
         if (!ImGui.Begin("Template Entity Editor")) return;
 
+        ImGui.Spacing();
+        ImGui.Spacing();
         if (ImGui.BeginTabBar("TemplateEditorTabs", ImGuiTabBarFlags.None))
         {
             blockTemplateEditor.Render();
@@ -46,5 +49,6 @@ public class TemplateEditorGui
         }
 
         ImGui.End();
+        ImGui.PopStyleVar();
     }
 }